requirements.txt
h5py
numpy
keras
scikit
learn
scipy
